Remove some signed/unsigned warnings

+++ SofaKernel/modules/SofaSimpleFem/TetrahedronDiffusionFEMForceField.inl
@@ -415,7 +415,7 @@ void TetrahedronDiffusionFEMForceField<DataTypes>::addKToMatrix(const core::Mech
     sofa::core::behavior::MultiMatrixAccessor::MatrixRef r = matrix->getMatrix(this->mstate);
     sofa::defaulttype::BaseMatrix* mat = r.matrix;
 
-    if((int)(mat->colSize()) != (m_topology->getNbPoints()*N) || (int)(mat->rowSize()) != (m_topology->getNbPoints()*N))
+    if((mat->colSize()) != (m_topology->getNbPoints()*N) || (mat->rowSize()) != (m_topology->getNbPoints()*N))
     {
         msg_error()<<"Wrong size of the input Matrix: need resize in addKToMatrix function.";
         mat->resize(m_topology->getNbPoints()*N,m_topology->getNbPoints()*N);
